requirements_total.txt
-e third_party/lerobot
absl-py==2.3.1
actionlib==1.14.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
angles==1.9.14
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
blinker==1.9.0
bondpy==1.8.7
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
datasets==3.6.0
deepdiff==8.5.0
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.9
docutils==0.22
draccus==0.10.0
dynamic-reconfigure==1.7.6
einops==0.8.1
evdev==1.9.2
Farama-Notifications==0.0.4
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.3.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
gitdb==4.0.12
GitPython==3.1.45
glfw==2.9.0
gnupg==2.3.1
grpcio==1.74.0
gym-aloha==0.1.1
gym-pusht==0.1.5
gymnasium==0.29.1
hf-xet==1.1.5
hf_transfer==0.1.9
huggingface-hub==0.34.3
hydra-core==1.3.2
idna==3.10
image-geometry==1.16.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inquirerpy==0.3.4
interactive-markers==1.12.2
itsdangerous==2.2.0
Jinja2==3.1.6
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
jsonlines==4.0.0
kornia==0.8.1
kornia_rs==0.1.9
labmaze==1.0.6
laser-geometry==1.6.8
lazy_loader==0.4
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
message-filters==1.17.4
mpmath==1.3.0
mujoco==2.3.7
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orderly-set==5.5.0
packaging==25.0
pandas==2.3.1
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pymunk==6.11.1
pynput==1.8.1
PyOpenGL==3.1.9
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-qt-binding==0.4.6
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
regex==2025.7.34
requests==2.32.4
rerun-sdk==0.22.1
resource-retriever==1.12.10
rich==14.1.0
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rviz==1.14.26
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
sensor-msgs==1.13.2
sentry-sdk==2.34.1
shapely==2.1.1
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
smmap==5.0.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
tifffile==2025.5.10
timm==1.0.19
toml==0.10.2
topic-tools==1.17.4
torch==2.7.1
torchcodec==0.5
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xacro==1.14.20
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
kuavo_humanoid_sdk
pympler
matplotlib
-e .